Test that non-hodler can not deposit funds
