+++ test/helpers/assertRevert.js
@@ -1,0 +1,11 @@
+async function assertRevert(promise) {
+  try {
+    await promise;
+    assert.fail('Expected revert not received');
+  } catch (error) {
+    const revertFound = error.message.search('revert') >= 0;
+    assert(revertFound, `Expected "revert", got ${error} instead`);
+  }
+};
+
+module.exports = assertRevert;
